# Conflicts:
#	plugin/pbr/impl/texture/gl_irradiance_cubemap.cpp
#	plugin/pbr/impl/texture/gl_irradiance_cubemap.h
#	plugin/pbr/impl/texture/gl_radiance_cubemap.cpp
#	plugin/pbr/impl/texture/gl_radiance_cubemap.h
#	src/renderer/base/texture.cpp
#	src/renderer/base/texture.h
#	src/renderer/opengl/base/gl_cubemap.cpp
#	src/renderer/opengl/base/gl_cubemap.h
#	src/renderer/opengl/base/gl_texture_2d.cpp
#	src/renderer/opengl/base/gl_texture_2d.h
#	src/renderer/opengl/es30/gl_resource/gl_dynamic_cubemap.cpp
#	src/renderer/opengl/es30/gl_resource/gl_dynamic_cubemap.h
#	src/renderer/opengl/util/gl_util.cpp
#	src/renderer/vulkan/base/vk_texture_2d.cpp
#	src/renderer/vulkan/base/vk_texture_2d.h

+++ src/renderer/opengl/util/gl_util.cpp
@@ -165,7 +165,7 @@ void GLUtil::renderCubemap(GraphicsContext& graphicsContext, uint32_t id, Render
     };
 
     const sp<opengl::GLPipeline> glPipeline = pipeline;
-    glUseProgram(glPipeline->id());
+    glUseProgram(static_cast<GLuint>(glPipeline->id()));
     glPipeline->getUniform("u_Projection").setUniformMatrix4fv(1, GL_FALSE, captureProjection.value());
     texture.upload(graphicsContext, nullptr);
